:recycle: More refactoring.
