+++ setup.cfg
@@ -37,9 +37,9 @@ include_package_data = True
 setup_requires = setuptools_scm[toml]>=6.0
 
 install_requires =
-    pabo
     fitsio>=1.1.8
     numpy>=1.20.0
+    pabo @ git+https://github.com/astrogewgaw/pabo@main#egg=pabo
 
 [options.packages.find]
 where=src
