+++ src/priwo/__init__.py
@@ -1,7 +1,3 @@
 """
 I/O for common pulsar data formats.
 """
-
-from priwo._version import version
-
-__all__ = ["version"]
