+++ src/priwo/misc/__init__.py
@@ -1,0 +1,3 @@
+"""
+R/W pulsar data from miscellaneous formats.
+"""
