--- src/priwo/others/__init__.py
@@ -1,15 +1,0 @@
-"""
-R/W pulsar data formats from miscellaneous sources.
-"""
-
-from priwo.others.par import readpar
-from priwo.others.par import writepar
-from priwo.others.polycos import readpolycos
-from priwo.others.polycos import writepolycos
-
-__all__ = [
-    "readpar",
-    "writepar",
-    "readpolycos",
-    "writepolycos",
-]
