+++ src/priwo/presto/__init__.py
@@ -1,27 +1,3 @@
 """
-R/W pulsar data formats used by the PRESTO package.
+R/W pulsar data from formats used by the PRESTO package.
 """
-
-from priwo.presto.inf import readinf
-from priwo.presto.dat import readdat
-from priwo.presto.fft import readfft
-from priwo.presto.bpf import readbpf
-from priwo.presto.pfd import readpfd
-from priwo.presto.inf import writeinf
-from priwo.presto.dat import writedat
-from priwo.presto.fft import writefft
-from priwo.presto.bpf import writebpf
-from priwo.presto.pfd import writepfd
-
-__all__ = [
-    "readinf",
-    "readdat",
-    "readfft",
-    "readbpf",
-    "readpfd",
-    "writeinf",
-    "writedat",
-    "writefft",
-    "writebpf",
-    "writepfd",
-]
