+++ src/priwo/timing/__init__.py
@@ -1,0 +1,3 @@
+"""
+R/W pulsar data from formats used by timing packages, like TEMPO/TEMPO2.
+"""
