+++ src/priwo/timing/coeff.py
