+++ src/priwo/timing/resid.py
